## Summary

Enable parallel testing for almost all tests in packages 'assert' and 'require' by calling t.Parallel() as the first line of the test.

A few tests are incompatible and will be fixed separately. They are marked with a FIXME.
Incompatible tests: TestFileExists, TestNoFileExists, TestDirExists, TestNoDirExists.

Before:
```console
  $ go test -count=10 ./assert ./require
  ok  	github.com/stretchr/testify/assert	7.575s
  ok  	github.com/stretchr/testify/require	1.501s
```

After:
```console
  $ go test -count=10 ./assert ./require
  ok  	github.com/stretchr/testify/assert	1.703s
  ok  	github.com/stretchr/testify/require	1.245s
```

## Changes
* `assert`: add `t.Parallel()` in almost all top level tests.
* `require`: add `t.Parallel()` in all top level tests.

## Motivation
The testsuite runs faster if GOMAXPROCS > 1.

## Related issues
* golang/go#73805